1) Aggiornata versione di karate e rimosso karate-apache in 	mvn/dependencies/testsuite/test/pom.xml 	lib/openspcoop2.userlibraries 	lib/openspcoop2-testsuite.userlibraries
2)

	Aggiunta dipendenza graalvm in sotto
		mvn/dependencies/testsuite/test/pom.xml
		lib/openspcoop2-testsuite.userlibraries

	E sotto il build.xml della testsuite di modipa

	<include name="js-21.2.0.jar"/>
  	<include name="js-scriptengine-21.2.0.jar"/>
	<include name="graal-sdk-21.2.0.jar"/>
	<include name="truffle-api-21.2.0.jar"/>
	<include name="icu4j-69.1.jar"/>

Restano da aggiornare le openspcoop2-testsuite.userlibraries

3)

Il FeatureServer è stato sostituito dal MockServer che utilizza il pattern del Builder per creare il mock,
quindi è diventata più malleabile e data oriented la costruzione dei mock server.

Questo:
	 File file = FileUtils.getFileRelativeTo(BloccanteRestTest.class, "modipa-mock.feature");
     FeatureServer server = FeatureServer.start(file, Integer.valueOf(prop.getProperty("http_port")), false, new HashMap<String,Object>((Map) prop));

Diventa Questo:

	  File file = ResourceUtils.getFileRelativeTo(BloccanteRestTest.class, "modipa-mock.feature");
      MockServer  server = MockServer
    			.feature(file)
    			.args(new HashMap<String,Object>((Map) prop))
    			.http(Integer.valueOf(prop.getProperty("http_port")))
    			.build();

4)

Per colpa di un bug di regressione ho dovuto modificare il src/org/openspcoop2/core/protocolli/modipa/testsuite/soap/sicurezza_messaggio/proxy.feature
e aggiungere il parametro response alle chiamate alla check_server_token

Lo stesso ho dovuto fare per il parametro body delle chiamate alla check_signature

5)

Un altro bug di regressione fa si che la karate.get(responseHeaders[header]) tratta lo header in maniera case sensitive, ho aperto una issue

+++ protocolli/modipa/testsuite/src/utils/get-id-collaborazione.js
@@ -9,7 +9,7 @@ function get_traccia(id_transazione) {
 
     db_sleep_before_read = karate.properties['db_sleep_before_read']
     
-    java.lang.Thread.sleep(db_sleep_before_read)
+    java.lang.Thread.sleep(java.lang.Integer.valueOf(db_sleep_before_read))
     DbUtils = Java.type('org.openspcoop2.core.protocolli.modipa.testsuite.DbUtils')
     db = new DbUtils(govwayDbConfig)
     dbquery = "select id_collaborazione from transazioni where id ='"+id_transazione+"'"
